example_CLIs/tronscan-cli2/pyproject.toml
[project]
name = "tronscan-cli"
version = "1.0.0"
description = "Python CLI client for Tronscan TRON blockchain explorer API"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pandas>=1.5.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "pylint>=2.17",
    "isort>=5.0",
]

[project.scripts]
tronscan = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
